pyproject.toml
[tool.poetry]
name = "repl_python3_bot-1"
version = "0.1.0"
description = ""
authors = ["Sebass1but1ab <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
numexpr = "^2.7.3"
aiohttp = "^3.7.4"
discord = "^1.7.3"
Flask = "^2.0.1"

dadjokes = "^1.3.2"
pyinsults = "^0.0.0"
PyNaCl = "^1.4.0"
ffmpeg = "^1.4"
mutagen = "^1.45.1"
topggpy = "^1.3.0"
waifu-py = "^1.0.3"
spacy = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
